Decouple Consensus II (move ethash-specific behavior)
